Move time utilities from CLI to Core
Moved time formatting utilities from src/cli/src/shared/reporting/time.ts to src/core/src/utils/time.ts

Rationale:
- The formatDuration, createVerboseDurationLogger, and timeSync functions are general-purpose time utilities with no CLI-specific dependencies
- They belong in Core alongside other utility modules (number, string, math, etc.) to be available across all packages
- Maintains proper architectural boundaries: CLI now imports these utilities from Core namespace rather than containing them

Changes:
- Created src/core/src/utils/time.ts with time utilities
- Updated src/core/src/utils/index.ts to export time module
- Updated src/cli/src/dependencies.ts to import from Core instead of local path
- Moved tests from src/cli/test/shared-time-utils.test.ts to src/core/test/time-utils.test.ts
- Removed old files from CLI package
- All tests pass with no new failures introduced
