+++ src/cli/src/shared/reporting/index.ts
@@ -1,2 +1,1 @@
 export * from "./byte-format.js";
-export * from "./time.js";
