+++ src/core/src/utils/index.ts
@@ -13,6 +13,7 @@ export * from "./numeric-options.js";
 export * from "./object.js";
 export * from "./regexp.js";
 export * from "./string.js";
+export * from "./time.js";
 export * from "./resolver-controller.js";
 export * from "./deprecated-builtin-variable-replacements.js";
 export * from "./readonly-view.js";
